scripts/nmsg_build.sh: install nmsgtool as 'sie-nmsgtool' and use private library directory
to avoid conflicting with another nmsgtool/libnmsg installed on the
system, we need to install nmsgtool to a path other than
/usr/bin/nmsgtool, and the libnmsg library and plugins need to be
installed into a private directory.

here we use '/usr/bin/sie-nmsgtool' for the nmsgtool utility,
'/usr/$lib/sie-dns-sensor' for the private library directory, and
'/usr/$lib/sie-dns-sensor/nmsg' for the private library's plugin
directory.

also, we don't actually need the 'libnmsg.so' symlink, so avoid copying
that symlink into the output directory. ("*.so*" -> "*.so.*")