debian: rename '/usr/bin/sie-nmsgtool' -> '/usr/lib/sie-dns-sensor/nmsgtool'
the argument/environment parsing code that nmsgtool uses [0] builds an
environment variable based on the process name, but it doesn't
canonicalize away "-", so we cannot actually use "sie-nmsgtool" as the
process name for the nmsgtool binary.  (that is, in order to call the
binary "sie-nmsgtool" we'd need to export a shell variable named
"ARGV_SIE-NMSGTOOL" which is not syntactically valid.)

so put the nmsgtool binary in /usr/lib/sie-dns-sensor/ without the
'sie-' prefix.  also use "sie-dns-sensor.pid" as the pidfile name.

[0] https://256.com/sources/argv/argv.html