debian/sie-dns-sensor.postrm: pass "--force" to deluser
it's possible for a laggard uploader to still be running for a few
seconds after sie-dns-sensor shuts down, which can prevent the removal
of the 'sie-dns-sensor' user. (userdel has a check that prevents users
from being removed if there are any processes owned by the user that are
still running.) pass "--force" to deluser in order to override this
check.