scripts/nmsg_build.sh: Use the system's libpcap
Since we now only support Debian 7 and newer, the system's libpcap is
new enough that we can avoid statically linking our own copy of libpcap.
This *may* help in situations where libpcap needs to be replaced at
runtime, e.g., in order to support Endace DAG cards.

This requires libpcap0.8-dev to be installed in the build chroot.